refactor: Move Dashboard component to a new directory structure refactor: Move Dashboard component to a new directory structure
- Updated import path for the Dashboard component to reflect new directory organization.
- Removed the old Dashboard component file as it has been relocated.
