+++ packages/Frontend/src/constants/chart.js
@@ -1,0 +1,36 @@
+export const DAYS_OF_WEEK = ["S", "M", "T", "W", "Th", "F", "Sa"];
+
+export const CHART_COLORS = {
+  light: {
+    axis: "#9CA3AF",
+    chartBg: "#fff",
+    line: "#111827",
+    dotFill: "#fff",
+    progress: "#4F46E5",
+    progressBg: "#1F2937"
+  },
+  dark: {
+    axis: "#ffffff",
+    chartBg: "#222222",
+    line: "#a5b4fc",
+    dotFill: "#1c2841",
+    progress: "#4F46E5",
+    progressBg: "#1F2937"
+  }
+};
+
+export const CHART_CONFIG = {
+  margin: { top: 20, right: 10, left: -20, bottom: 10 },
+  dot: {
+    r: 6,
+    stroke: "#4F46E5",
+    strokeWidth: 3,
+    filter: "drop-shadow(0 2px 6px #4F46E533)"
+  },
+  activeDot: {
+    r: 8,
+    fill: "#4F46E5",
+    strokeWidth: 3,
+    filter: "drop-shadow(0 2px 6px #4F46E533)"
+  }
+};
